AS-IS) startNum int -> Long TO-BE) 채팅 GroupBy 기능 고도화

+++ src/main/java/com/spring/sharepod/v1/repository/SearchForm.java
@@ -20,7 +20,7 @@ public class SearchForm {
     @Length(max = 100)
     private String boardRegion;
 
-    private int startNum;
+    private Long startNum;
 
     private String searchTitle;
 }
